==> src/0_lsms_processing/0_download_country_codes.ipynb <==
# Download Countries Code

This code will download the country code aligned with the ISO used by the World Bank. The alignment is important, since we will use the LSMS collected by the world bank and with the correct codes, we can autotomize some parts.  We load the list of all Sub-Saharan Africa from [Wikipedia](https://en.wikipedia.org/wiki/Sub-Saharan_Africa) and get the country codes from the [World Bank API](https://microdata.worldbank.org/api-documentation/catalog/index.html#operation/listCountryCodes). Since the name are not always same, I did the work and added the missing countries by hand.

import pandas as pd
import requests


df: pd.DataFrame = pd.read_csv("../../data/countries_meta/countries.csv")
countries: set[str] = set(df["country"])

url: str = "https://microdata.worldbank.org/index.php/api/catalog/country_codes"
codes: any = requests.get(url).json()

def diff(l1: list[any], l2: list[any]) -> list[any]:
	"""Get difference between two lists. 

	Args:
		l1 (list[any]): list 1
		l2 (list[any]): list 2

	Returns:
		A list of the different values. 
	"""
	return list(set(l1) - set(l2)) + list(set(l2) - set(l1))

countries_code: dict[list[str], list[str]] = {
    "name": [],
    "iso": []
}

verify: list[str] = []
for code in codes["country_codes"]:
    if code["name"] in countries:
        countries_code["name"].append(code["name"])
        countries_code["iso"].append(code["iso"])
        verify.append(code["name"])

diff(verify, df["country"]) # output for manual work

countries_code["name"].append("Republic of the Congo")
countries_code["iso"].append("COG")
countries_code["name"].append("Democratic Republic of the Congo")
countries_code["iso"].append("COD")
countries_code["name"].append("Ivory Coast")
countries_code["iso"].append("CI")
countries_code["name"].append("Cabo Verde")
countries_code["iso"].append("CPV")

pd.DataFrame.from_dict(countries_code).to_csv("../../data/countries_meta/countries_code.csv", index=False)
